🎨 Serve worker.js as static files
The WashU browser needs worker.js for rendering tracks such as bigWig.
Previously the file is served via proxying to the API server. However,
the logic isn't quite sound since worker.js is solely for client-side
rendering, and has nothing to do with the API.

Now worker.js is prefetched and saved in /public and will be served with
other static assets.